test(ai-sdk): colocate tool-output-denied converter tests with helpers.ts
Move the converter coverage from __tests__/tool-call-approval.test.ts to
helpers.test.ts per the test colocation rule (CodeRabbit review).
