+++ client-sdks/ai-sdk/src/helpers.test.ts
@@ -263,3 +263,41 @@ describe('finish usage conversion', () => {
     });
   });
 });
+
+describe('tool-output-denied chunk conversion (issue #20880)', () => {
+  const chunk = {
+    type: 'tool-output-denied' as const,
+    runId: 'run-123',
+    from: ChunkFrom.AGENT,
+    payload: {
+      toolCallId: 'tooluse_abc123',
+      toolName: 'myTool',
+      args: { param: 'value' },
+      approval: { id: 'approval-1', approved: false as const, reason: 'Not allowed' },
+    },
+  };
+
+  it('converts the Mastra denial to an AI SDK v6 stream part', () => {
+    expect(convertMastraChunkToAISDKv6({ chunk, mode: 'stream' })).toEqual({
+      type: 'tool-output-denied',
+      toolCallId: 'tooluse_abc123',
+      toolName: 'myTool',
+    });
+  });
+
+  it('converts the stream part to an AI SDK UI message chunk', () => {
+    expect(
+      convertFullStreamChunkToUIMessageStream({
+        part: {
+          type: 'tool-output-denied',
+          toolCallId: 'tooluse_abc123',
+          toolName: 'myTool',
+        },
+        onError: String,
+      }),
+    ).toEqual({
+      type: 'tool-output-denied',
+      toolCallId: 'tooluse_abc123',
+    });
+  });
+});
